requirements.txt
streamlit>=1.32.0
tensorflow>=2.20.0
numpy<2
pillow
opencv-python-headless